Consolidated orderbook implementations into a single package.
