+++ src/main/scala-2.11/markets/matching/FindFirstMatchingFunction.scala
@@ -17,7 +17,7 @@ package markets.matching
 
 import java.util.UUID
 
-import markets.generic
+import markets.orderbooks
 import markets.tradables.orders.{Order, Predicate}
 
 
@@ -37,7 +37,7 @@ class FindFirstMatchingFunction[O1 <: Order, O2 <: Order with Predicate[O1]] ext
     *       contained in the `orderBook`.  Depending on the type of `orderBook`, the result of this `MatchingFunction`
     *       may be non-deterministic.
     */
-  def apply(order: O2, orderBook: generic.OrderBook[O1, collection.GenMap[UUID, O1]]): Option[(O2, O1)] = {
+  def apply(order: O2, orderBook: orderbooks.OrderBook[O1, collection.GenMap[UUID, O1]]): Option[(O2, O1)] = {
     orderBook.find(order.isAcceptable) match {
       case Some(matchingOrder) => Some(order, matchingOrder)
       case None => None
