+++ src/main/scala-2.11/markets/orderbooks/mutable/OrderBook.scala
@@ -13,44 +13,42 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package markets.parallel.concurrent.orderbooks
+package markets.orderbooks.mutable
 
 import java.util.UUID
 
-import markets.generic
-import markets.tradables.orders.Order
+import markets.orderbooks
 import markets.tradables.Tradable
+import markets.tradables.orders.Order
 
-import scala.collection.parallel
+import scala.collection.generic.CanBuildFrom
+import scala.collection.mutable
 
 
-/** Class for modeling an `OrderBook` for use when thread-safe access is required.
+/** Class for modeling an `OrderBook`.
   *
   * @param tradable all `Orders` contained in the `OrderBook` should be for the same `Tradable`.
-  * @tparam O type of `Order` stored in the `OrderBook`.
-  * @todo Currently the underlying `existingOrders` will use the JVM default ForkJoinTaskSupport object for scheduling
-  *       and load-balancing.  This [[http://docs.scala-lang.org/overviews/parallel-collections/configuration.html can be customized]]
-  *       but requires some clear thinking about how to expose this functionality to the user.
+  * @tparam O type of `Order` stored in the order book.
+  * @tparam CC type of underlying collection class used to store the `Order` instances.
   */
-class OrderBook[O <: Order](val tradable: Tradable) extends generic.OrderBook[O, parallel.immutable.ParMap[UUID, O]] {
+class OrderBook[O <: Order, +CC <: mutable.Map[UUID, O]](val tradable: Tradable)(implicit cbf: CanBuildFrom[_, _, CC])
+  extends orderbooks.OrderBook[O, CC] {
 
   /** Add an `Order` to the `OrderBook`.
     *
     * @param order the `Order` that should be added to the `OrderBook`.
-    * @note adding an `Order` to the `OrderBook` is an `O(1)` operation.
     */
   def add(order: O): Unit = {
     require(order.tradable == tradable)
-    existingOrders.synchronized { existingOrders = existingOrders + (order.uuid -> order) }
+    existingOrders += (order.uuid -> order)
   }
 
   /** Filter the `OrderBook` and return those `Order` instances satisfying the given predicate.
     *
     * @param p predicate defining desirable `Order` characteristics.
     * @return collection of `Order` instances satisfying the given predicate.
-    * @note filtering the `OrderBook` is an `O(n)` operation.
     */
-  def filter(p: (O) => Boolean): Option[parallel.ParIterable[O]] = {
+  def filter(p: (O) => Boolean): Option[Iterable[O]] = {
     val filteredOrders = existingOrders.values.filter(p)
     if (filteredOrders.isEmpty) None else Some(filteredOrders)
   }
@@ -59,7 +57,6 @@ class OrderBook[O <: Order](val tradable: Tradable) extends generic.OrderBook[O,
     *
     * @param p predicate defining desirable `Order` characteristics.
     * @return `None` if no `Order` in the `OrderBook` satisfies the predicate; `Some(order)` otherwise.
-    * @note finding an `Order` in the `OrderBook` is an `O(n)` operation.
     */
   def find(p: (O) => Boolean): Option[O] = existingOrders.values.find(p)
 
@@ -80,31 +77,41 @@ class OrderBook[O <: Order](val tradable: Tradable) extends generic.OrderBook[O,
 
   /** Remove and return an existing `Order` from the `OrderBook`.
     *
-    * @param uuid the `UUID` for the `Order` that should be removed from the `OrderBook`.
-    * @return `None` if the `uuid` is not found in the `OrderBook`; `Some(order)` otherwise.
-    * @note removing and returning an `Order` from the `OrderBook` is an `O(1)` operation.
+    * @param uuid the `UUID` for the order that should be removed from the `OrderBook`.
+    * @return `None` if the `uuid` is not found in the order book; `Some(order)` otherwise.
     */
-  def remove(uuid: UUID): Option[O] = existingOrders.synchronized {
-    existingOrders.get(uuid) match {
-      case residualOrder@Some(order) => existingOrders = existingOrders - uuid; residualOrder
-      case None => None
-    }
-  }
+  def remove(uuid: UUID): Option[O] = existingOrders.remove(uuid)
 
-  /* Protected at package-level for testing; volatile for thread-safety. */
-  @volatile protected[orderbooks] var existingOrders = parallel.immutable.ParMap.empty[UUID, O]
+  /* Protected at package-level for testing. */
+  protected[orderbooks] val existingOrders: CC = cbf().result()
 
 }
 
 
-/** Factory for creating `OrderBook` instances. */
+/** Companion object for `OrderBook`.
+  *
+  * Used as a factory for creating `OrderBook` instances.
+  */
 object OrderBook {
 
-  /** Create a `OrderBook` instance for a particular `Tradable`.
+  /** Create an `OrderBook` instance for a particular `Tradable`.
     *
     * @param tradable all `Orders` contained in the `OrderBook` should be for the same `Tradable`.
-    * @tparam O type of `Order` stored in the `OrderBook`.
+    * @tparam O type of `Order` stored in the order book.
+    * @tparam CC type of underlying collection class used to store the `Order` instances.
     */
-  def apply[O <: Order](tradable: Tradable): OrderBook[O] = new OrderBook[O](tradable)
+  def apply[O <: Order, CC <: mutable.Map[UUID, O]](tradable: Tradable)(implicit cbf: CanBuildFrom[_, _, CC]): OrderBook[O, CC] =  {
+    new OrderBook[O, CC](tradable)(cbf)
+  }
+
+  /** Create an `OrderBook` instance for a particular `Tradable`.
+    *
+    * @param tradable all `Orders` contained in the `OrderBook` should be for the same `Tradable`.
+    * @tparam O type of `Order` stored in the order book.
+    */
+  def apply[O <: Order](tradable: Tradable): OrderBook[O, mutable.HashMap[UUID, O]] =  {
+    val cbf = implicitly[CanBuildFrom[_, _, mutable.HashMap[UUID, O]]]
+    new OrderBook[O, mutable.HashMap[UUID, O]](tradable)(cbf)
+  }
 
 }
