+++ src/main/scala-2.11/markets/orderbooks/mutable/SortedOrderBook.scala
@@ -13,13 +13,13 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package markets.mutable.orderbooks
+package markets.orderbooks.mutable
 
 import java.util.UUID
 
-import markets.generic
-import markets.tradables.orders.Order
+import markets.orderbooks
 import markets.tradables.Tradable
+import markets.tradables.orders.Order
 
 import scala.collection.generic.CanBuildFrom
 import scala.collection.mutable
@@ -35,7 +35,7 @@ import scala.collection.mutable
   */
 class SortedOrderBook[O <: Order, +CC <: mutable.Map[UUID, O]](val tradable: Tradable)
                                                              (implicit ordering: Ordering[O], cbf: CanBuildFrom[_, _, CC])
-  extends generic.OrderBook[O, CC] with generic.SortedOrders[O, CC, mutable.TreeSet[O]] {
+  extends orderbooks.OrderBook[O, CC] with orderbooks.SortedOrders[O, CC, mutable.TreeSet[O]] {
 
   /** Add an `Order` to the `SortedOrderBook`.
     *
