+++ src/main/scala-2.11/markets/orderbooks/parallel/concurrent/OrderBook.scala
@@ -13,23 +13,26 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package markets.concurrent.orderbooks
+package markets.orderbooks.parallel.concurrent
 
 import java.util.UUID
 
-import markets.generic
-import markets.tradables.orders.Order
+import markets.orderbooks
 import markets.tradables.Tradable
+import markets.tradables.orders.Order
 
-import scala.collection.immutable
+import scala.collection.parallel
 
 
 /** Class for modeling an `OrderBook` for use when thread-safe access is required.
   *
   * @param tradable all `Orders` contained in the `OrderBook` should be for the same `Tradable`.
   * @tparam O type of `Order` stored in the `OrderBook`.
+  * @todo Currently the underlying `existingOrders` will use the JVM default ForkJoinTaskSupport object for scheduling
+  *       and load-balancing.  This [[http://docs.scala-lang.org/overviews/parallel-collections/configuration.html can be customized]]
+  *       but requires some clear thinking about how to expose this functionality to the user.
   */
-class OrderBook[O <: Order](val tradable: Tradable) extends generic.OrderBook[O, immutable.Map[UUID, O]] {
+class OrderBook[O <: Order](val tradable: Tradable) extends orderbooks.OrderBook[O, parallel.immutable.ParMap[UUID, O]] {
 
   /** Add an `Order` to the `OrderBook`.
     *
@@ -47,7 +50,7 @@ class OrderBook[O <: Order](val tradable: Tradable) extends generic.OrderBook[O,
     * @return collection of `Order` instances satisfying the given predicate.
     * @note filtering the `OrderBook` is an `O(n)` operation.
     */
-  def filter(p: (O) => Boolean): Option[Iterable[O]] = {
+  def filter(p: (O) => Boolean): Option[parallel.ParIterable[O]] = {
     val filteredOrders = existingOrders.values.filter(p)
     if (filteredOrders.isEmpty) None else Some(filteredOrders)
   }
@@ -63,14 +66,12 @@ class OrderBook[O <: Order](val tradable: Tradable) extends generic.OrderBook[O,
   /** Return the head `Order` of the `OrderBook`.
     *
     * @return `None` if the `OrderBook` is empty; `Some(order)` otherwise.
-    * @note returning the head `Order` of the `OrderBook` is an `O(1)` operation.
     */
   def headOption: Option[O] = existingOrders.values.headOption
 
   /** Remove and return the head `Order` of the `OrderBook`.
     *
     * @return `None` if the `OrderBook` is empty; `Some(order)` otherwise.
-    * @note removing and returning the head `Order` of the `OrderBook` is an `O(1)` operation.
     */
   def remove(): Option[O] = headOption match {
     case Some(order) => remove(order.uuid)
@@ -85,21 +86,18 @@ class OrderBook[O <: Order](val tradable: Tradable) extends generic.OrderBook[O,
     */
   def remove(uuid: UUID): Option[O] = existingOrders.synchronized {
     existingOrders.get(uuid) match {
-      case residualOrder @ Some(order) => existingOrders = existingOrders - uuid; residualOrder
+      case residualOrder@Some(order) => existingOrders = existingOrders - uuid; residualOrder
       case None => None
     }
   }
 
   /* Protected at package-level for testing; volatile for thread-safety. */
-  @volatile protected[orderbooks] var existingOrders = immutable.Map.empty[UUID, O]
+  @volatile protected[orderbooks] var existingOrders = parallel.immutable.ParMap.empty[UUID, O]
 
 }
 
 
-/** Companion object for `OrderBook`.
-  *
-  * Used as a factory for creating `OrderBook` instances.
-  */
+/** Factory for creating `OrderBook` instances. */
 object OrderBook {
 
   /** Create a `OrderBook` instance for a particular `Tradable`.
