+++ src/main/scala-2.11/markets/orderbooks/parallel/mutable/package.scala
@@ -13,10 +13,9 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package markets.parallel.mutable
-
+package markets.orderbooks.parallel
 
 /** Classes providing parallel implementations of order books that have been optimized for performance in a single-threaded
   * simulation. Classes in this package are not thread-safe.
   */
-package object orderbooks
+package object mutable
