+++ src/main/scala-2.11/markets/orderbooks/parallel/package.scala
@@ -13,10 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package markets.concurrent
+package markets.orderbooks
 
 
-/** Classes providing implementations of order books that have been optimized for performance in a multi-threaded
-  * simulation. Classes in this package are guaranteed thread-safe.
-  */
-package object orderbooks
+package object parallel
